+++ NEWS.rst
@@ -1,5 +1,22 @@
             pyAggr3g470r project news
 
+6.4: 2015-03-17
+    New:
+     * a new page 'history' to explore your database of articles.
+    Changes:
+     * updated documentation;
+     * minor improvements.
+    Fixes:
+     * changed the type of the column 'last_modified' to string.
+
+6.3: 2015-03-08
+    New:
+    * a new architecture with base for controllers;
+    * new, more complete RESTful API;
+    * a crawler handling errors and cache control;
+    * the new crawler takes advantage of the RESTful API
+    (can be run on the client side).
+
 6.2: 2015-02-26
     The system of email notifications for new articles has been removed.
     This feature was hardly used.